added loader class in addition to registry provider
