+++ presidio-analyzer/presidio_analyzer/recognizer_registry/recognizer_registry_provider.py
@@ -2,14 +2,14 @@ from __future__ import annotations
 
 import logging
 import re
-from collections.abc import ItemsView
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Type, Union
+from typing import Dict, Optional, Union
 
-import yaml
-
-from presidio_analyzer import EntityRecognizer, PatternRecognizer
 from presidio_analyzer.recognizer_registry import RecognizerRegistry
+from presidio_analyzer.recognizer_registry.recognizers_loader_utils import (
+    RecognizerConfigurationLoader,
+    RecognizerListLoader,
+)
 
 logger = logging.getLogger("presidio-analyzer")
 
@@ -48,210 +48,16 @@ class RecognizerRegistryProvider:
         conf_file: Optional[Union[Path, str]] = None,
         registry_configuration: Optional[Dict] = None,
     ):
-        self.configuration = self._get_configuration(
+        self.configuration = RecognizerConfigurationLoader.get(
             conf_file=conf_file, registry_configuration=registry_configuration
         )
-        self.supported_languages = None
-        self.all_existing_recognizers = self._get_all_existing_recognizers()
         return
-
-    @staticmethod
-    def _add_missing_keys(configuration: Dict, conf_file: Union[Path, str]) -> Dict:
-        """
-        Add missing keys to the configuration.
-
-        Missing keys are added using the default configuration read from file.
-
-        :param configuration: The configuration to update.
-        :param conf_file: The configuration file to read from.
-        """
-
-        defaults = yaml.safe_load(open(conf_file))
-        configuration.update(
-            {k: v for k, v in defaults.items() if k not in list(configuration.keys())}
-        )
-        return configuration
-
-    def _get_configuration(
-        self, conf_file: Union[Path, str], registry_configuration: Dict
-    ) -> Union[Dict[str, Any]]:
-        """Get the configuration from the provided file or dict.
-
-        :param conf_file: The configuration file
-        to read the recognizer registry configuration from.
-        :param registry_configuration: The configuration to use.
-        """
-
-        if conf_file and registry_configuration:
-            raise ValueError(
-                "Either conf_file or registry_configuration should"
-                " be provided, not both."
-            )
-
-        configuration = {}
-
-        if registry_configuration:
-            configuration = registry_configuration.copy()
-
-        if not conf_file:
-            configuration = self._add_missing_keys(
-                configuration=configuration, conf_file=self._get_full_conf_path()
-            )
-        else:
-            try:
-                configuration = self._add_missing_keys(
-                    configuration=configuration, conf_file=conf_file
-                )
-            except OSError:
-                logger.warning(
-                    f"configuration file {conf_file} not found.  "
-                    f"Using default config."
-                )
-                configuration = self._add_missing_keys(
-                    configuration=configuration, conf_file=self._get_full_conf_path()
-                )
-            except Exception:
-                logger.warning(
-                    f"Failed to parse file {conf_file}, " f"resorting to default"
-                )
-                configuration = self._add_missing_keys(
-                    configuration=configuration, conf_file=self._get_full_conf_path()
-                )
-
-        return configuration
-
-    @staticmethod
-    def _is_recognizer_enabled(recognizer_conf: Union[Dict[str, Any], str]) -> bool:
-        return "enabled" not in recognizer_conf or recognizer_conf["enabled"]
-
-    @staticmethod
-    def _get_recognizer_name(recognizer_conf: Union[Dict[str, Any], str]) -> str:
-        if isinstance(recognizer_conf, str):
-            return recognizer_conf
-        return recognizer_conf["name"]
-
-    @staticmethod
-    def _get_recognizer_context(
-        recognizer: Union[Dict[str, Any], str],
-    ) -> Optional[List[str]]:
-        if isinstance(recognizer, str):
-            return None
-        return recognizer.get("context", None)
-
-    @staticmethod
-    def _get_recognizer_items(
-        recognizer_conf: Union[Dict[str, Any], str],
-    ) -> Union[dict[Any, Any], ItemsView[str, Any]]:
-        if isinstance(recognizer_conf, str):
-            return {}
-        return recognizer_conf.items()
-
-    def _get_recognizer_languages(
-        self, recognizer_conf: Union[Dict[str, Any], str]
-    ) -> List[Dict[str, Any]]:
-        """
-        Get the different language properties for each recognizer.
-
-        Creating a new recognizer for each supported language.
-        If language wasn't specified, create a recognizer for each supported language.
-
-        :param recognizer_conf: The aforementioned recognizer.
-        :return: The list of recognizers in the supported languages.
-        """
-        if (
-            isinstance(recognizer_conf, str)
-            or "supported_languages" not in recognizer_conf
-        ):
-            return [
-                {
-                    "supported_language": language,
-                    "context": self._get_recognizer_context(recognizer_conf),
-                }
-                for language in self.supported_languages
-            ]
-
-        if isinstance(recognizer_conf["supported_languages"][0], str):
-            return [
-                {"supported_language": language, "context": None}
-                for language in recognizer_conf["supported_languages"]
-            ]
-
-        return [
-            {
-                "supported_language": language["language"],
-                "context": language.get("context", None),
-            }
-            for language in recognizer_conf["supported_languages"]
-        ]
-
-    @staticmethod
-    def _split_recognizers(
-        recognizers_conf: Union[Dict[str, Any], str],
-    ) -> Tuple[List[Union[str, Dict[str, Any]]], List[Union[str, Dict[str, Any]]]]:
-        """
-        Split the recognizer list to predefined and custom.
-
-        All recognizers are custom by default though
-        type: 'custom' can be mentioned as well.
-        This function supports the previous format as well.
-
-        :param recognizers_conf: The recognizers' configuration
-        """
-
-        predefined = [
-            recognizer_conf
-            for recognizer_conf in recognizers_conf
-            if ("type" in recognizer_conf and recognizer_conf["type"] == "predefined")
-        ]
-        custom = [
-            recognizer_conf
-            for recognizer_conf in recognizers_conf
-            if not isinstance(recognizer_conf, str)
-            and ("type" not in recognizer_conf or recognizer_conf["type"] == "custom")
-        ]
-        return predefined, custom
-
-    def _create_custom_recognizers(
-        self, recognizer_conf: Dict
-    ) -> List[PatternRecognizer]:
-        """Create a custom recognizer for each language, based on the provided conf."""
-        # legacy recognizer
-        if "supported_language" in recognizer_conf:
-            return [PatternRecognizer.from_dict(recognizer_conf)]
-
-        recognizers = []
-
-        for supported_language in self._get_recognizer_languages(recognizer_conf):
-            copied_recognizer = {
-                k: v
-                for k, v in recognizer_conf.items()
-                if k not in ["enabled", "type", "supported_languages"]
-            }
-            kwargs = {**copied_recognizer, **supported_language}
-            recognizers.append(PatternRecognizer.from_dict(kwargs))
-
-        return recognizers
 
     default_values = {
         "supported_languages": ["en"],
         "recognizers": [],
         "global_regex_flags": re.DOTALL | re.MULTILINE | re.IGNORECASE,
     }
-
-    def _is_language_supported_globally(
-            self,
-            recognizer: EntityRecognizer) -> bool:
-        if recognizer.supported_language not in self.supported_languages:
-            logger.warning(
-                f"Recognizer not added to registry because "
-                f"language is not supported by registry - "
-                f"{recognizer.name} supported "
-                f"languages: {recognizer.supported_language}"
-                f", registry supported languages: "
-                f"{', '.join(self.supported_languages)}"
-            )
-            return False
-        return True
 
     def create_recognizer_registry(self) -> RecognizerRegistry:
         """Create a recognizer registry according to configuration loaded previously."""
@@ -269,107 +75,8 @@ class RecognizerRegistryProvider:
                 )
             fields[field] = self.configuration.get(field, self.default_values[field])
 
-        fields["recognizers"] = self.init_recognizers(fields["recognizers"],
-                                                      fields["supported_languages"],
-                                                      fields["global_regex_flags"])
+        fields["recognizers"] = RecognizerListLoader.get(fields["recognizers"],
+                                                                  fields["supported_languages"],
+                                                                  fields["global_regex_flags"])
 
         return RecognizerRegistry(**fields)
-
-    def init_recognizers(
-            self,
-            recognizers: Dict[str, Any],
-            supported_languages: Iterable[str],
-            global_regex_flags: int,
-        ) -> Iterable[EntityRecognizer]:
-        """
-        Create an iterator of recognizers.
-
-        The recognizers are initialized according to configuration loaded previously.
-        """
-        self.supported_languages = supported_languages
-        self.global_regex_flags = global_regex_flags
-        recognizer_instances = []
-        predefined, custom = self._split_recognizers(recognizers)
-        for recognizer_conf in predefined:
-            for language_conf in self._get_recognizer_languages(recognizer_conf):
-                if self._is_recognizer_enabled(recognizer_conf):
-                    copied_recognizer_conf = {
-                        k: v
-                        for k, v in self._get_recognizer_items(recognizer_conf)
-                        if k not in ["enabled", "type", "supported_languages", "name"]
-                    }
-                    kwargs = {**copied_recognizer_conf, **language_conf}
-                    recognizer_name = self._get_recognizer_name(recognizer_conf)
-                    recognizer_cls = self._get_existing_recognizer_cls_by_name(
-                        recognizer_name
-                    )
-                    recognizer_instances.append(recognizer_cls(**kwargs))
-
-        for recognizer_conf in custom:
-            if self._is_recognizer_enabled(recognizer_conf):
-                recognizer_instances.extend(
-                    self._create_custom_recognizers(recognizer_conf)
-                )
-
-        for recognizer_conf in recognizer_instances:
-            if isinstance(recognizer_conf, PatternRecognizer):
-                recognizer_conf.global_regex_flags = self.global_regex_flags
-
-
-        recognizer_instances = [
-            recognizer
-            for recognizer in recognizer_instances
-            if self._is_language_supported_globally(recognizer)
-        ]
-
-        return recognizer_instances
-
-
-    @staticmethod
-    def _get_full_conf_path(
-        default_conf_file: Union[Path, str] = "default_recognizers.yaml",
-    ) -> Path:
-        """Return a Path to the default conf file."""
-        return Path(Path(__file__).parent, "../conf", default_conf_file)
-
-    @staticmethod
-    def _get_all_existing_recognizers(
-        cls: Optional[Type[EntityRecognizer]] = None,
-    ) -> Set[Type[EntityRecognizer]]:
-        """
-        Return all subclasses of EntityRecognizer, recursively.
-
-        :param cls: The initial class, if None, cls=EntityRecognizer.
-        """
-
-        if not cls:
-            cls = EntityRecognizer
-
-        return set(cls.__subclasses__()).union(
-            [
-                s
-                for c in cls.__subclasses__()
-                for s in RecognizerRegistryProvider._get_all_existing_recognizers(c)
-            ]
-        )
-
-    def _get_existing_recognizer_cls_by_name(
-        self, recognizer_name: str
-    ) -> Type[EntityRecognizer]:
-        """
-        Get the recognizer class by name.
-
-        Returns the requested recognizer class out of the list of all existing
-        recognizers currently inheriting from EntityRecognizer.
-        Raises a ValueError if the recognizer is not found.
-
-        :param recognizer_name: The name of the recognizer.
-        """
-        for recognizer in self.all_existing_recognizers:
-            if recognizer_name == recognizer.__name__:
-                return recognizer
-
-        raise ValueError(
-            f"Recognizer of name {recognizer_name} was not found in the "
-            f"list of recognizers inheriting the EntityRecognizer class"
-        )
